feat(Category): utiliza repositório no UpdateCategoryHandler
O `UpdateCategoryHandler` foi atualizado para utilizar um padrão de
repositório, removendo a dependência direta do `DimaDbContext` e
tornando o código mais testável.

O construtor agora aceita `ICategoryRepository` e `IUnitOfWork`
em vez de `DimaDbContext`, e o método para recuperar uma categoria pelo
identificador foi atualizado para chamar `GetByIdAsync` do repositório.

Adicionalmente, o método `UpdateCategory` agora utiliza o repositório
para buscar a categoria e atualizá-la, e o método `SaveChangesAsync` do
`IUnitOfWork` para persistência. A interface `ICategoryRepository` foi
estendida a fim de corresponder à funcionalidade.

BREAKING CHANGE: Construtor e método UpdateCategoryHandler alterados.

O construtor agora requer uma instância de `ICategoryRepository` e
`IUnitOfWork` para funcionar corretamente. Atualize o código existente
para prover essas dependências.
