+++ src/Dima.Core/Categories/ICategoryRepository.cs
@@ -1,8 +1,7 @@
 namespace Dima.Core.Categories;
 
 /// <summary>
-/// <para>Interface for the Category repository.</para>
-/// <para>Provides methods to interact with the Category data store.</para>
+/// Provides methods to interact with the Category data store.
 /// </summary>
 public interface ICategoryRepository
 {
@@ -13,4 +12,18 @@ public interface ICategoryRepository
     /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
     /// <returns>A task that represents the asynchronous operation.</returns>
     Task AddAsync(Category category, CancellationToken cancellationToken = default);
+
+    /// <summary>
+    /// Asynchronously retrieves a category by its identifier.
+    /// </summary>
+    /// <param name="id">The identifier of the category to retrieve.</param>
+    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
+    /// <returns>A task that represents the asynchronous operation. The task result contains the category if found; otherwise, null.</returns>
+    Task<Category> GetByIdAsync(long id, CancellationToken cancellationToken = default);
+
+    /// <summary>
+    /// Updates an existing category in the data store.
+    /// </summary>
+    /// <param name="category">The category to update.</param>
+    void Update(Category category);
 }
